@@ -4,30 +4,32 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@dynamic-devs/symbol-angular": "^0.0.14",
+    "@angular/animations": "14.0.0",
+    "@angular/cdk": "14.0.0",
+    "@angular/common": "14.0.0",
+    "@angular/compiler": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/forms": "14.0.0",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
+    "@dynamic-devs/symbol": "^0.1.2",
+    "@dynamic-devs/symbol-angular": "^0.1.2",
     "@nestjs/common": "^8.0.0",
     "@nestjs/config": "^2.0.1",
     "@nestjs/core": "^8.0.0",
     "@nestjs/platform-express": "^8.0.0",
-    "@nrwl/angular": "14.1.9",
+    "@nrwl/angular": "14.2.4",
     "googleapis": "^100.0.0",
     "joi": "^17.6.0",
+    "ngx-sharebuttons": "^11.0.0",
     "pdf-creator-node": "^2.3.4",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~7.4.0",
@@ -35,40 +37,40 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.0",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
+    "@angular-devkit/build-angular": "14.0.0",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.0",
+    "@angular/language-service": "14.0.0",
     "@nestjs/schematics": "^8.0.0",
     "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "14.1.9",
-    "@nrwl/cypress": "14.1.9",
-    "@nrwl/eslint-plugin-nx": "14.1.9",
-    "@nrwl/jest": "14.1.9",
-    "@nrwl/linter": "14.1.9",
-    "@nrwl/nest": "14.1.9",
-    "@nrwl/node": "14.1.9",
-    "@nrwl/workspace": "14.1.9",
+    "@nrwl/cli": "14.2.4",
+    "@nrwl/cypress": "14.2.4",
+    "@nrwl/eslint-plugin-nx": "14.2.4",
+    "@nrwl/jest": "14.2.4",
+    "@nrwl/linter": "14.2.4",
+    "@nrwl/nest": "14.2.4",
+    "@nrwl/node": "14.2.4",
+    "@nrwl/workspace": "14.2.4",
     "@tailwindcss/line-clamp": "^0.4.0",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "autoprefixer": "^10.4.7",
     "cypress": "^9.1.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "nx": "14.1.9",
+    "jest-preset-angular": "11.1.2",
+    "nx": "14.2.4",
     "postcss": "^8.4.14",
-    "prettier": "^2.5.1",
+    "prettier": "2.6.2",
     "tailwindcss": "^3.0.24",
     "ts-jest": "27.1.4",
-    "typescript": "~4.6.2"
+    "typescript": "4.7.3"
   }
 }
