@@ -9,9 +9,16 @@ import { AppRoutingModule } from './app.routing';
 import { CertificateComponent } from './certificate/certificate.component';
 import { SymbolsModule } from '@dynamic-devs/symbol-angular';
 
+import { ShareModule } from 'ngx-sharebuttons';
+
+const LIBRARIES = [SymbolsModule, ShareModule];
+
 @NgModule({
   declarations: [AppComponent, CertificateComponent, NxWelcomeComponent],
-  imports: [BrowserModule, HttpClientModule, ElementsModule, AppRoutingModule, SymbolsModule],
+  imports: [
+    BrowserModule, HttpClientModule, ElementsModule, AppRoutingModule,
+    ...LIBRARIES
+  ],
   providers: [],
   bootstrap: [AppComponent],
 })
